Delete old files.
